sdk/python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fractary-forge"
version = "1.0.0"
description = "Python SDK for Fractary Forge - Create, manage, and execute AI agent definitions"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Fractary", email = "hello@fractary.com"}
]
keywords = ["ai", "agents", "llm", "fractary", "forge", "sdk"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "httpx>=0.25.0",
    "semver>=3.0.0",
]

[project.optional-dependencies]
langchain = [
    "langchain>=0.1.0",
    "langchain-anthropic>=0.1.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "types-PyYAML",
]

[project.urls]
Homepage = "https://github.com/fractary/forge"
Documentation = "https://github.com/fractary/forge#readme"
Repository = "https://github.com/fractary/forge"
Issues = "https://github.com/fractary/forge/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["fractary_forge*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"
asyncio_mode = "auto"
